tests: move some functions to helpers
